pyproject.toml
[tool.poetry]
name = "presetify"
version = "0.1.0"
description = "Extract Lightroom presets from image metadata and generate XMP files"
authors = ["JayLindblad"]
readme = "README.md"
packages = [{include = "presetify", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
textual = "^0.82.0"
pillow = "^10.4.0"
httpx = "^0.27.0"
pyexiftool = "^0.5.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
black = "^24.8.0"
ruff = "^0.6.0"

[tool.poetry.scripts]
presetify = "presetify.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
